backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==11.3.0
numpy==1.24.3
easyocr==1.7.0
python-jose[cryptography]==3.3.0
tesseract==0.1.3
pytesseract==0.3.13
